Move firebase user functions to utils folder
